Move get_traced_jetstream from utils to telemetry
